+++ resources/forms-ui/src/app/app.module.ts
@@ -131,7 +131,8 @@ import { ExecViewSuperExectivesPageComponent } from './pages/dashboard/executive
 import { ExecViewBranchExectivesPageComponent } from './pages/dashboard/executive/exec-view-branch-exectives-page/exec-view-branch-exectives-page.component';
 import { ExecViewFrontDesksPageComponent } from './pages/dashboard/executive/exec-view-front-desks-page/exec-view-front-desks-page.component';
 import { FormsCountryPickerComponent } from './components/forms-country-picker/forms-country-picker.component';
-import { ClientFormsDraftsPageComponent } from './pages/dashboard/client/client-forms-drafts-page/client-forms-drafts-page.component';
+import { ClientFavoriteFormsPageComponent } from './pages/dashboard/client/client-favorite-forms-page/client-favorite-forms-page.component';
+import { ClientSuggestMerchantPageComponent } from './pages/dashboard/client/client-suggest-merchant-page/client-suggest-merchant-page.component';
 
 @NgModule({
   declarations: [
@@ -247,7 +248,8 @@ import { ClientFormsDraftsPageComponent } from './pages/dashboard/client/client-
     ExecViewBranchExectivesPageComponent,
     ExecViewFrontDesksPageComponent,
     FormsCountryPickerComponent,
-    ClientFormsDraftsPageComponent,
+    ClientFavoriteFormsPageComponent,
+    ClientSuggestMerchantPageComponent,
   ],
   imports: [
     BrowserModule,