+++ resources/forms-ui/src/app/pages/dashboard/client/client-form-new-entry-page/client-form-new-entry-page.component.ts
@@ -6,7 +6,9 @@ import { ClipboardService } from 'ngx-clipboard';
 import { Users } from 'src/app/models/users.model';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { BranchService } from 'src/app/services/branch/branch.service';
 import { ClientService } from 'src/app/services/client/client.service';
+import { CompanyBranches } from 'src/app/models/company-branches.model';
 import { Component, OnInit, ViewChild, TemplateRef } from '@angular/core';
 import { EndpointService } from 'src/app/services/endpoint/endpoint.service';
 import { ReloadingService } from 'src/app/services/reloader/reloading.service';
@@ -25,8 +27,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
   user: Users;
   imgUrl: string;
   pinCode: string;
+  status: number;
   loading: boolean;
+  saved: boolean;
   created: boolean;
+  branch_id: number;
   hasFile: boolean;
   formFiles: number;
   formInstance: any;
@@ -40,17 +45,22 @@ export class ClientFormNewEntryPageComponent implements OnInit {
   pinRequired: boolean;
   updateProfile: boolean;
   submissionCode: string;
+  loadingBranches: boolean;
   showAttachments: boolean;
   docDialogRef: NgbModalRef;
   pinDialogRef: NgbModalRef;
+  disableValidation: boolean;
   loadingAttachments: boolean;
   setPinDialogRef: NgbModalRef;
   attachmentFiles: Array<File>;
   attachmentKeys: Array<string>;
   existingAttachments: Array<any>;
+  selectBranchDialogRef: NgbModalRef;
+  branchesList: Array<CompanyBranches>;
   @ViewChild('pin', { static: false }) pinDialog: TemplateRef<any>;
   @ViewChild('setPin', { static: false }) setPinDialog: TemplateRef<any>;
   @ViewChild('confirm', { static: false }) confirmDialog: TemplateRef<any>;
+  @ViewChild('selectBranch', { static: false }) selectBranchDialog: TemplateRef<any>;
   @ViewChild('viewImgAttachment', { static: false }) viewImgDialog: TemplateRef<any>;
   @ViewChild('viewDocAttachment', { static: false }) viewDocDialog: TemplateRef<any>;
 
@@ -61,18 +71,23 @@ export class ClientFormNewEntryPageComponent implements OnInit {
     private reloader: ReloadingService,
     private clipboard: ClipboardService,
     private clientService: ClientService,
+    private branchesService: BranchService,
     private formBuilder: FormBuilderService,
     private endpointService: EndpointService,
     private localStorage: LocalStorageService,
     private downloadService: DownloaderService,
     private fileUploadService: FileUploadsService
   ) {
+    this.status = 0;
     this.pinCode = '';
     this.formFiles = 0;
+    this.branch_id = null;
+    this.branchesList = [];
     this.submissionCode = '';
     this.attachmentKeys = [];
     this.attachmentFiles = [];
     this.existingAttachments = [];
+    this.disableValidation = false;
     this.form = history.state.form;
     this.form = this.reloader.resolveDataLoss(this.form);
 
@@ -81,6 +96,7 @@ export class ClientFormNewEntryPageComponent implements OnInit {
     console.log('submission_code: ' + this.form.submission_code);
     this.getFormAttachments(this.user.id.toString());
     this.checkIfUserHasFormPin();
+    this.generateSubmissionCode();
   }
 
   ngOnInit() {
@@ -99,7 +115,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
   }
 
   generateSubmissionCode() {
-    return Math.random().toString(36).substr(2, 5);
+    this.clientService.generateFormSubmissionCode().then(
+      code => {
+        this.submissionCode = code;
+      }
+    );
   }
 
   resolveStrCharacters(e: KeyboardEvent) {
@@ -154,8 +174,14 @@ export class ClientFormNewEntryPageComponent implements OnInit {
   setFormData(data: any) {
     this.clientService.getDetails(this.user.id.toString()).then(
       client => {
-        this.clientProfile = client.client_details[0];
-        this.clientService.autoFillFormData(data, this.clientProfile);
+        if (client.length != 0) {
+          this.clientProfile = client.client_details[0];
+          this.clientService.autoFillFormData(data, this.clientProfile);
+        }
+        else {
+          this.clientProfile = this.localStorage.getUser();
+          console.log('_____clientProfile: ' + JSON.stringify(this.clientProfile));
+        }
       },
       error => {
         console.log('error: ' + JSON.stringify(error));
@@ -165,6 +191,26 @@ export class ClientFormNewEntryPageComponent implements OnInit {
 
   getFormData() {
     return this.formInstance.userData;
+  }
+
+  getBranches() {
+    this.loadingBranches = true;
+    this.branchesService.getCompanyBranches(this.form.merchant_id).then(
+      branches => {
+        this.branchesList = branches;
+        this.loadingBranches = false;
+      },
+      error => {
+        this.loadingBranches = false;
+        console.log('error loading branches');
+      }
+    );
+  }
+
+  chooseBranch(branch_id: number) {
+    this.branch_id = branch_id;
+    console.log('am closing');
+    this.selectBranchDialogRef.close();
   }
 
   appendOnChangeEventToFileInput() {
@@ -236,8 +282,8 @@ export class ClientFormNewEntryPageComponent implements OnInit {
 
   submitFormAndAttachments(user_data: any, updateProfile: boolean) {
     console.log('is submitting');
-    const form_submission_code = this.generateSubmissionCode();
-    this.submissionCode = form_submission_code;
+    const form_submission_code = this.submissionCode;
+    // this.submissionCode = form_submission_code;
     if (this.hasFile) {
       this.uploadFormAttachments(user_data, updateProfile, form_submission_code);
     }
@@ -245,11 +291,12 @@ export class ClientFormNewEntryPageComponent implements OnInit {
       const update = updateProfile ? 1 : 0;
       const filled_data = this.formBuilder.getFormUserData(user_data);
       const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-      this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+      console.log('_____updated_data: ' + JSON.stringify(updated_data));
+      this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branch_id).then(
         ok => {
           if (ok) {
             this.loading = false;
-            this.created = true;
+            this.status == 0 ? this.created = true : this.saved = true;
           }
           else {
             this.loading = false;
@@ -263,10 +310,15 @@ export class ClientFormNewEntryPageComponent implements OnInit {
   }
 
   submitForm() {
-    this.loading = true;
     const user_data = this.getFormData();
     console.log(JSON.stringify(user_data));
     console.log('this form: ' + this.formBuilder.getFormUserData(user_data));
+    !this.disableValidation
+      ? this.submitFormWithValidation(user_data)
+      : this.submitFormWithoutValidation(user_data);
+  }
+
+  submitFormWithValidation(user_data: any) {
     const unfilled = this.clientService.validateFormFilled(user_data);
     console.log('unfilled: ' + JSON.stringify(unfilled));
     if (unfilled.length != 0) {
@@ -285,14 +337,40 @@ export class ClientFormNewEntryPageComponent implements OnInit {
     }
   }
 
+  submitFormWithoutValidation(user_data: any) {
+    this.submitFormAndAttachments(user_data, this.updateProfile);
+  }
+
   submit() {
     this.modalService.open(this.confirmDialog, { centered: true }).result.then(
       result => {
         if (result == 'yes') {
-          this.handlePinCode(true);
+          if (this.form.can_view == 0) {
+            this.showMerchantBranchesDialog(true);
+          }
+          else {
+            this.handlePinCode(true);
+          }
         }
         else {
-          this.handlePinCode(false);
+          if (this.form.can_view == 0) {
+            this.showMerchantBranchesDialog(false);
+          }
+          else {
+            this.handlePinCode(false);
+          }
+        }
+      }
+    );
+  }
+
+  showMerchantBranchesDialog(update: boolean) {
+    this.getBranches();
+    this.selectBranchDialogRef = this.modalService.open(this.selectBranchDialog, { centered: true });
+    this.selectBranchDialogRef.result.then(
+      result => {
+        if (!_.isNull(result) || !_.isUndefined(result)) {
+          this.handlePinCode(update);
         }
       }
     );
@@ -417,11 +495,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
         const update = updateProfile ? 1 : 0;
         const filled_data = this.formBuilder.getFormUserData(user_data);
         const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-        this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code).then(
+        this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code, this.status, this.branch_id).then(
           ok => {
             if (ok) {
               this.loading = false;
-              this.created = true;
+              this.status == 0 ? this.created = true : this.saved = true;
             }
             else {
               this.loading = false;
@@ -445,11 +523,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
           const update = updateProfile ? 1 : 0;
           const filled_data = this.formBuilder.getFormUserData(user_data);
           const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branch_id).then(
             _ok => {
               if (_ok) {
                 this.loading = false;
-                this.created = true;
+                this.status == 0 ? this.created = true : this.saved = true;
               }
               else {
                 this.loading = false;
@@ -495,11 +573,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
           const update = updateProfile ? 1 : 0;
           const filled_data = this.formBuilder.getFormUserData(user_data);
           const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branch_id).then(
             ok => {
               if (ok) {
                 this.loading = false;
-                this.created = true;
+                this.status == 0 ? this.created = true : this.saved = true;
               }
               else {
                 this.loading = false;
@@ -524,11 +602,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
           const update = updateProfile ? 1 : 0;
           const filled_data = this.formBuilder.getFormUserData(user_data);
           const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branch_id).then(
             _ok => {
               if (_ok) {
                 this.loading = false;
-                this.created = true;
+                this.status == 0 ? this.created = true : this.saved = true;
               }
               else {
                 this.loading = false;
@@ -591,11 +669,11 @@ export class ClientFormNewEntryPageComponent implements OnInit {
           const update = updateProfile ? 1 : 0;
           const filled_data = this.formBuilder.getFormUserData(user_data);
           const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code).then(
+          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code, this.status, this.branch_id).then(
             ok => {
               if (ok) {
                 this.loading = false;
-                this.created = true;
+                this.status == 0 ? this.created = true : this.saved = true;
               }
               else {
                 this.loading = false;
@@ -691,6 +769,12 @@ export class ClientFormNewEntryPageComponent implements OnInit {
     e.stopPropagation();
     this.documentUrl = this.endpointService.apiHost + 'storage/attachments/' + url;
     this.docDialogRef = this.modalService.open(this.viewDocDialog, { centered: true });
+  }
+
+  saveAsDraft() {
+    this.status = 4;
+    this.disableValidation = true;
+    this.handlePinCode(false);
   }
 
   copy() {