+++ resources/forms-ui/src/app/pages/dashboard/client/client-forms-entry-page/client-forms-entry-page.component.ts
@@ -6,7 +6,9 @@ import { ClipboardService } from 'ngx-clipboard';
 import { Users } from 'src/app/models/users.model';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { BranchService } from 'src/app/services/branch/branch.service';
 import { ClientService } from 'src/app/services/client/client.service';
+import { CompanyBranches } from 'src/app/models/company-branches.model';
 import { Component, OnInit, ViewChild, TemplateRef } from '@angular/core';
 import { EndpointService } from 'src/app/services/endpoint/endpoint.service';
 import { ReloadingService } from 'src/app/services/reloader/reloading.service';
@@ -14,6 +16,7 @@ import { DownloaderService } from 'src/app/services/downloader/downloader.servic
 import { LocalStorageService } from 'src/app/services/storage/local-storage.service';
 import { FormBuilderService } from 'src/app/services/form-builder/form-builder.service';
 import { FileUploadsService } from 'src/app/services/file-uploads/file-uploads.service';
+import { CompanyService } from 'src/app/services/company/company.service';
 
 @Component({
   selector: 'app-client-forms-entry-page',
@@ -25,11 +28,14 @@ export class ClientFormsEntryPageComponent implements OnInit {
   form: any;
   user: Users;
   imgUrl: string;
+  status: number;
   pinCode: string;
   loading: boolean;
+  saved: boolean;
   created: boolean;
   hasFile: boolean;
   formFiles: number;
+  branchId: number;
   formInstance: any;
   formRenderer: any;
   clientProfile: any;
@@ -41,6 +47,7 @@ export class ClientFormsEntryPageComponent implements OnInit {
   pinRequired: boolean;
   updateProfile: boolean;
   submissionCode: string;
+  loadingBranches: boolean;
   showAttachments: boolean;
   docDialogRef: NgbModalRef;
   pinDialogRef: NgbModalRef;
@@ -49,9 +56,12 @@ export class ClientFormsEntryPageComponent implements OnInit {
   attachmentFiles: Array<File>;
   attachmentKeys: Array<string>;
   existingAttachments: Array<any>;
+  selectBranchDialogRef: NgbModalRef;
+  branchesList: Array<CompanyBranches>;
   @ViewChild('pin', { static: false }) pinDialog: TemplateRef<any>;
   @ViewChild('setPin', { static: false }) setPinDialog: TemplateRef<any>;
   @ViewChild('confirm', { static: false }) confirmDialog: TemplateRef<any>;
+  @ViewChild('selectBranch', { static: false }) selectBranchDialog: TemplateRef<any>;
   @ViewChild('viewImgAttachment', { static: false }) viewImgDialog: TemplateRef<any>;
   @ViewChild('viewDocAttachment', { static: false }) viewDocDialog: TemplateRef<any>;
 
@@ -62,14 +72,19 @@ export class ClientFormsEntryPageComponent implements OnInit {
     private reloader: ReloadingService,
     private clipboard: ClipboardService,
     private clientService: ClientService,
+    private branchesService: BranchService,
+    private companyService: CompanyService,
     private formBuilder: FormBuilderService,
     private endpointService: EndpointService,
     private localStorage: LocalStorageService,
     private downloadService: DownloaderService,
     private fileUploadService: FileUploadsService
   ) {
+    this.status = 0;
     this.pinCode = '';
     this.formFiles = 0;
+    this.branchId = null;
+    this.branchesList = [];
     this.submissionCode = '';
     this.attachmentKeys = [];
     this.attachmentFiles = [];
@@ -84,6 +99,7 @@ export class ClientFormsEntryPageComponent implements OnInit {
       ? this.getFormAttachments(this.form.submission_code)
       : this.getAttachmentsForCurrentForm(this.user.id.toString());
     this.checkIfUserHasFormPin();
+    this.generateSubmissionCode();
   }
 
   ngOnInit() {
@@ -102,7 +118,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
   }
 
   generateSubmissionCode() {
-    return Math.random().toString(36).substr(2, 5);
+    this.clientService.generateFormSubmissionCode().then(
+      code => {
+        this.submissionCode = code;
+      }
+    );
   }
 
   resolveStrCharacters(e: KeyboardEvent) {
@@ -161,6 +181,25 @@ export class ClientFormsEntryPageComponent implements OnInit {
 
   getFormData() {
     return this.formInstance.userData;
+  }
+
+  getBranches() {
+    this.loadingBranches = true;
+    this.branchesService.getCompanyBranches(this.form.merchant_id).then(
+      branches => {
+        this.branchesList = branches;
+        this.loadingBranches = false;
+      },
+      error => {
+        this.loadingBranches = false;
+        console.log('error loading branches');
+      }
+    );
+  }
+
+  chooseBranch(branch_id: number) {
+    this.branchId = branch_id;
+    this.selectBranchDialogRef.close();
   }
 
   appendOnChangeEventToFileInput() {
@@ -232,8 +271,8 @@ export class ClientFormsEntryPageComponent implements OnInit {
 
   submitFormAndAttachments(user_data: any, updateProfile: boolean) {
     console.log('is submitting');
-    const form_submission_code = this.generateSubmissionCode();
-    this.submissionCode = form_submission_code;
+    const form_submission_code = this.submissionCode;
+    alert('code: ' + form_submission_code);
     if (this.hasFile) {
       this.uploadFormAttachments(user_data, updateProfile, form_submission_code);
     }
@@ -241,11 +280,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
       const update = updateProfile ? 1 : 0;
       const filled_data = this.formBuilder.getFormUserData(user_data);
       const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-      this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+      this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branchId).then(
         ok => {
           if (ok) {
             this.loading = false;
-            this.created = true;
+            this.status == 0 ? this.created = true : this.saved = true;
           }
           else {
             this.loading = false;
@@ -256,6 +295,18 @@ export class ClientFormsEntryPageComponent implements OnInit {
         }
       );
     }
+  }
+
+  showMerchantBranchesDialog(update: boolean) {
+    this.getBranches();
+    this.selectBranchDialogRef = this.modalService.open(this.selectBranchDialog, { centered: true });
+    this.selectBranchDialogRef.result.then(
+      result => {
+        if (!_.isNull(result) || !_.isUndefined(result)) {
+          this.handlePinCode(update);
+        }
+      }
+    );
   }
 
   handlePinCode(update: boolean) {
@@ -296,10 +347,20 @@ export class ClientFormsEntryPageComponent implements OnInit {
     this.modalService.open(this.confirmDialog, { centered: true }).result.then(
       result => {
         if (result == 'yes') {
-          this.handlePinCode(true);
+          if (this.form.can_view == 0) {
+            this.showMerchantBranchesDialog(true);
+          }
+          else {
+            this.handlePinCode(true);
+          }
         }
         else {
-          this.handlePinCode(false);
+          if (this.form.can_view == 0) {
+            this.showMerchantBranchesDialog(false);
+          }
+          else {
+            this.handlePinCode(false);
+          }
         }
       }
     );
@@ -403,11 +464,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
               const update = updateProfile ? 1 : 0;
               const filled_data = this.formBuilder.getFormUserData(user_data);
               const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-              this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+              this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branchId).then(
                 _ok => {
                   if (_ok) {
                     this.loading = false;
-                    this.created = true;
+                    this.status == 0 ? this.created = true : this.saved = true;
                   }
                   else {
                     this.loading = false;
@@ -452,11 +513,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
               const update = updateProfile ? 1 : 0;
               const filled_data = this.formBuilder.getFormUserData(user_data);
               const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-              this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+              this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branchId).then(
                 ok => {
                   if (ok) {
                     this.loading = false;
-                    this.created = true;
+                    this.status == 0 ? this.created = true : this.saved = true;
                   }
                   else {
                     this.loading = false;
@@ -482,11 +543,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
             const update = updateProfile ? 1 : 0;
             const filled_data = this.formBuilder.getFormUserData(user_data);
             const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-            this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code).then(
+            this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, form_submission_code, this.status, this.branchId).then(
               _ok => {
                 if (_ok) {
                   this.loading = false;
-                  this.created = true;
+                  this.status == 0 ? this.created = true : this.saved = true;
                 }
                 else {
                   this.loading = false;
@@ -498,8 +559,6 @@ export class ClientFormsEntryPageComponent implements OnInit {
                 console.log('form submission error 4');
               }
             );
-            // this.created = true;
-            // this.loading = false;
           }
           else {
             console.log('file upload failed');
@@ -534,11 +593,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
         const update = updateProfile ? 1 : 0;
         const filled_data = this.formBuilder.getFormUserData(user_data);
         const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-        this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code).then(
+        this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code, this.status, this.branchId).then(
           ok => {
             if (ok) {
               this.loading = false;
-              this.created = true;
+              this.status == 0 ? this.created = true : this.saved = true;
             }
             else {
               this.loading = false;
@@ -550,8 +609,6 @@ export class ClientFormsEntryPageComponent implements OnInit {
             console.log('form submission error 5');
           }
         );
-        // this.created = true;
-        // this.loading = false;
       }
     });
   }
@@ -579,11 +636,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
           const update = updateProfile ? 1 : 0;
           const filled_data = this.formBuilder.getFormUserData(user_data);
           const updated_data = this.clientService.getUpdatedClientFormData(JSON.parse(filled_data), this.clientProfile);
-          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code).then(
+          this.clientService.submitForm(_.toString(this.user.id), this.form.form_code, this.clientProfile, JSON.parse(updated_data), update, submission_code, this.status, this.branchId).then(
             ok => {
               if (ok) {
                 this.loading = false;
-                this.created = true;
+                this.status == 0 ? this.created = true : this.saved = true;
               }
               else {
                 this.loading = false;
@@ -704,6 +761,11 @@ export class ClientFormsEntryPageComponent implements OnInit {
     e.stopPropagation();
     this.documentUrl = this.endpointService.apiHost + 'storage/attachments/' + url;
     this.docDialogRef = this.modalService.open(this.viewDocDialog, { centered: true });
+  }
+
+  saveAsDraft() {
+    this.status = 4;
+    this.handlePinCode(false);
   }
 
   copy() {