+++ resources/forms-ui/src/app/pages/dashboard/user-account-creator/user-account-creator.component.ts
@@ -312,7 +312,7 @@ export class UserAccountCreatorComponent implements OnInit {
       ? merchant_id
       : this.getSelectedMerchantIdentifier();
     console.log('selected company_id: ' + id);
-    this.branchService.getBranch(id.toString()).then(
+    this.branchService.getCompanyBranches(id.toString()).then(
       res => {
         const branches = res as any;
         console.log('res_length: ' + branches.length);