+++ resources/forms-ui/src/app/pages/dashboard/user-account-editor/user-account-editor.component.ts
@@ -247,7 +247,7 @@ export class UserAccountEditorComponent implements OnInit {
     this.branchNamesList = [];
     const id = this.getSelectedMerchantIdentifier();
     console.log('selected company_id: ' + id);
-    this.branchService.getBranch(id.toString()).then(
+    this.branchService.getCompanyBranches(id.toString()).then(
       res => {
         const branches = res as any;
         console.log('res_length: ' + branches.length);