+++ resources/forms-ui/src/app/services/date-time/date-time.service.ts
@@ -40,4 +40,16 @@ export class DateTimeService {
   safeDateFormat(date: string) {
     return date.replace(/-/g, '/');
   }
+
+  bootstrapDateFormat(date: string) {
+    // Bootstrap date picker returns single digit for months from Jan to Sept
+    // In order to allow us to compare against MYSQL which returns double digits
+    // for that, we convert the month accordingly.
+    const _date = date as any;
+    const _day = _.toNumber(_date.day) <= 9 ? '0' + _date.day : _date.day;
+    const _month = _.toNumber(_date.month) <= 9 ? '0' + _date.month : _date.month;
+
+    const final_date = _date.year + '-' + _month + '-' + _day;
+    return final_date;
+  }
 }